+++ modules/core/services.nix
@@ -4,7 +4,15 @@
     libinput.enable = true; # Input Handling
     fstrim.enable = true; # SSD Optimizer
     gvfs.enable = true; # For Mounting USB & More
-    openssh.enable = true; # Enable SSH
+    openssh = {
+      enable = true; # Enable SSH
+      settings = {
+        PermiRootLogin = "no"; # Prevent root from SSH login
+        PasswordAuthentication = true; #Users can SSH using kb and password
+        KbdInteractiveAuthentication = true;
+      };
+      ports = [22];
+    };
     blueman.enable = true; # Bluetooth Support
     tumbler.enable = true; # Image/video preview
     gnome.gnome-keyring.enable = true;